2008-01-19  Stefan Jahn  <stefan@lkcc.org>
        * log_amp.cpp (log_amp): Added new Logarithmic Amplifier verilog
        device.  Thanks to Mike Brinson!

+++ qucs/qucs/diagrams/truthdiagram.cpp
@@ -111,7 +111,7 @@ if(g) if(!g->cPointsX.isEmpty()) {
     colWidth = metrics.width("0");
     if(z > colWidth)  colWidth = z;
     colWidth += 2;
-    counting = int(log(NumAll) / log(2) + 0.9999); // number of bits
+    counting = int(log(double(NumAll)) / log(2.0) + 0.9999); // number of bits
 
     if((x+colWidth*counting) >= x2) {    // enough space for text ?
       checkColumnWidth("0", metrics, 0, x2, y);
